:bug: Database split
I've split up the 'entries' table inside the database into 'entries' and 'users'. The table 'entries' now only has the UserID instead of the name for data redunction. However, I have some mistake in my code, because the username does not get catched and does not get displayed. The table row where the username should be displayed on the contributions page is empty.